% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneNames.R
\name{getGeneNames}
\alias{getGeneNames}
\title{Replaces Ensembl IDs with gene names}
\usage{
getGeneNames(species = c("", "Human", "Mouse"), x)
}
\arguments{
\item{species}{Organism. "Human","Mouse" datasets exist in package.}

\item{x}{The vector of Ensembl IDs to be replaced with the corresponding}
}
\value{
An equally-sized vector with replaced values, where possible.
}
\description{
Tries to replace the Ensembl IDs ("ENSGxxxxx.x") with gene names, using
internally stored data.
}
