% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackageData.R
\name{getPackageData}
\alias{getPackageData}
\title{Returns internal package datasets}
\usage{
getPackageData(
  species = c("", "Human", "Mouse"),
  what = c("TF", "miRNA", "GT")
)
}
\arguments{
\item{species}{Which organism to return data for. Everything else except "Human" or "Mouse" returns NULL.}

\item{what}{\code{TF} for reference dataset, \code{miRNA} for microRNA or \code{GT} for gene table}
}
\value{
Requested dataset or NULL
}
\description{
Returns internal package datasets
}
