% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_reference_from_files.R
\name{produce_reference_from_files}
\alias{produce_reference_from_files}
\title{Produce a list of Transcription Factors interactions}
\usage{
produce_reference_from_files(
  pi_file,
  tf_file,
  tf_column,
  bm_file,
  out_file,
  headers = rep(T, 3),
  seps = rep("", 3),
  ...
)
}
\arguments{
\item{pi_file}{Protein interactions from STRING (fully detailed, with header)}

\item{tf_file}{Transcription factors (Tab separated, with header)}

\item{tf_column}{Column containing gene name of TFs}

\item{bm_file}{Protein-gene matching from Biomart (N:1, comma separated, with header)}

\item{out_file}{Results will be saved to this tab-separated file}

\item{headers}{header parameter for \code{\link[utils:read.table]{utils::read.table()}} for each of the 3 input files}

\item{seps}{sep parameter for \code{\link[utils:read.table]{utils::read.table()}} for each of the 3 input files}

\item{...}{Arguments to be passed to \code{\link[=produce_reference]{produce_reference()}}}
}
\value{
Result of \code{\link[=produce_reference]{produce_reference()}} or NULL in case of error
}
\description{
Produces and saves a file with a list of Transcription Factors interactions,
by combining information from provided files#'
}
