% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNFct.R
\name{RNCircle}
\alias{RNCircle}
\title{Locate circles in regulatory network}
\usage{
RNCircle(network, verbose = F, throwOnError = T)
}
\arguments{
\item{network}{Two-column gene matrix. Every row contains one regulating and one regulated gene.}

\item{verbose}{Shows some information during process.}

\item{throwOnError}{If FALSE, in case of stopping error, returns NULL. Otherwise \code{\link[base:stop]{base::stop()}} is called.}
}
\value{
3-column matrix of circles. Column X regulates column Y, Y regulates Z
and Z regulates X.
}
\description{
Locate circles in regulatory network.
}
