% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNEA_v3.R
\name{RNEAv3}
\alias{RNEAv3}
\title{RNEA version 3}
\usage{
RNEAv3(
  filename,
  identifier = "GeneName",
  species,
  FC_threshold = 1,
  PV_threshold = 0.05,
  network = "regulatory",
  output_dir = ".",
  output = "Output",
  type_of_output = "csv"
)
}
\arguments{
\item{type_of_output}{}
}
\value{

}
\description{
RNEA version 3
}
