% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNRank.R
\docType{package}
\name{RNRank-package}
\alias{RNRank-package}
\alias{_PACKAGE}
\title{RNRank: Gene Ranking In Inferred Active Regulatory Network}
\description{
Infers active regulatory networks from gene expression data using a combination of prior knowledge and enrichment analysis and ranks the genes in them.
}
\author{
\strong{Maintainer}: Konstantinos Tragkas \email{std507586@ac.eap.gr} (\href{https://orcid.org/0000-0001-6181-8904}{ORCID})

}
\keyword{internal}
