% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNEA_v3.R
\name{RNEAv3}
\alias{RNEAv3}
\title{Regulatory Network Enrichment Analysis (v3)}
\usage{
RNEAv3(
  filename,
  species,
  internal_data = T,
  miRNA_data = F,
  FC_threshold = 1,
  PV_threshold = 0.05,
  output_dir = ".",
  output = "Output",
  reference_dir = "ReferenceFiles",
  ffl = T,
  circles = T,
  rank = T,
  favorFFL = T,
  favorCircles = T,
  ...,
  verbose = F,
  throwOnError = T
)
}
\arguments{
\item{filename}{Differential expression file (Cuffdiff, 3-column, 1-column)}

\item{species}{Organism. "Human","Mouse" datasets exist in package.
If any other, the relevant reference file must be provided in \code{reference_dir}.}

\item{internal_data}{If \code{TRUE} the package's internal datasets are used}

\item{miRNA_data}{Use miRNA data}

\item{FC_threshold}{log2FC threshold}

\item{PV_threshold}{p-value threshold}

\item{output_dir}{Directory to save output to}

\item{output}{Prefix for output files}

\item{reference_dir}{Directory containing reference files for unsupported species}

\item{ffl}{Locate and save Feed-Forward Loops by calling \code{\link[=RNFfl]{RNFfl()}}}

\item{circles}{Locate and save circles by calling \code{\link[=RNCircle]{RNCircle()}}}

\item{rank}{Call RNRank() on the resulting network and save an extra file with ranks}

\item{favorFFL}{If ranking, favor genes participating in FFLs}

\item{favorCircles}{If ranking, favor genes participating in circles}

\item{...}{Arguments to be passed to \code{\link[=RNRank]{RNRank()}}}

\item{verbose}{Shows some information during process.}

\item{throwOnError}{If FALSE, in case of stopping error, returns NULL. Otherwise \code{\link[base:stop]{base::stop()}} is called.}
}
\value{
"Analysis completed!"
}
\description{
Infers active regulatory networks from gene expression data using a combination
of prior knowledge and enrichment analysis
}
\details{
Transcription factors (TF) control a multitude of constitutive, cell-specific,
developmental, proliferative or homeostatic processes in the cells of all known
organisms. Due to their central role in gene regulation a considerable number
of human diseases have been associated with TF function, including metabolic,
autoimmune disorders and cancer. Despite the progress over the last years,
towards the identification of TF gene targets, either with experimental or
in silico approaches, we are still far from accurately reconstructing the
hierarchy of transcriptional regulators from genome-wide data. In this work
we are trying to overcome the existing limitations for the inference and study
of regulatory networks of TF-interactions and their subsequent use in
enriching our understanding of key biological processes. Given a differential
expression file RNEA extracts lists of prioritized TFs, and possibly miRNAs.
Most importantly a regulatory subnetwork is extracted showing how the activated
regulators interact with their target genes and between each other.
In this way, given a genome-wide expression experiment, RNEA prioritizes
important regulatory components.
It also succeeds to reconstruct meaningful subnetworks of gene regulation,
offering further ways of analyzing the data based on network theory.
A detailed guide describing RNEA's use can be found \href{https://sites.google.com/a/fleming.gr/rnea/manual}{here}.
}
