% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNRank.R
\name{RNRank}
\alias{RNRank}
\title{Rank by gene importance}
\usage{
RNRank(
  network,
  damping = 0.85,
  max_iterations = 100,
  threshold = 0,
  self = T,
  letZeros = F,
  divider = 100,
  sorted = T,
  preorder = F,
  favoredFFLs = NULL,
  favoredCircles = NULL,
  favorings = 1.1,
  verbose = F,
  throwOnError = T
)
}
\arguments{
\item{network}{Two-column gene matrix. Every row contains one regulating and one regulated gene.}

\item{damping}{Damping factor (0-1) defining probability of non-randomness.}

\item{max_iterations}{Maximum number of iterations, if not converging earlier (minimum 10).}

\item{threshold}{Euclidean distance between iterations less than or equal to this value
will terminate calculations}

\item{self}{Self regulations permitted (\code{TRUE}) or not (\code{FALSE}).}

\item{letZeros}{Set to \code{FALSE} to allow random regulations from regulating genes
to those not normally regulated.}

\item{divider}{Divides the probability used for dangling genes,
to use it for random regulations from regulating genes (minimum 10).}

\item{sorted}{Results sorted by descending probability (T) or unsorted (F).}

\item{preorder}{Sort input matrix, so that regulating genes with more targets are first.}

\item{favoredFFLs}{Regulations or genes, favored for better ranking. See 'Details'.}

\item{favoredCircles}{Regulations or genes, favored for better ranking. See 'Details'.}

\item{favorings}{Multiplier of favored regulations. Single value or numeric vector.
In case of vector, first value is for \code{favoredFFLs} and second for \code{favoredCircles}.}

\item{verbose}{Shows some information during process.}

\item{throwOnError}{If FALSE, in case of stopping error, returns NULL. Otherwise \code{\link[base:stop]{base::stop()}} is called.}
}
\value{
Named 1-column matrix of probabilities (0-1).
}
\description{
Infers importance of Regulatory Network nodes (genes),
by using a Google's PageRank-like algorithm.
}
\details{
\code{favoredFFLs} and \code{favoredCircles} contain gene names and are treated
differently, depending on them being vectors or matrices and if any of them
is a matrix, depending on the number of columns. In any other case, they are ignored.
\itemize{
\item Vector or 1-column matrix: The regulations with any other gene are favored.
\item 2 columns: Considered to contain directed edges. Exactly these regulations are favored.
\item More columns: \code{favoredFFLs} is considered to contain FFLs in the first three columns
(as the matrix returned from \code{\link[=RNFfl]{RNFfl()}}). The regulations C1->C2, C1->C3 and
C2->C3 are favored. \code{favoredCircles} is considered to contain circles in the
first three columns (as the matrix returned from \code{\link[=RNCircle]{RNCircle()}}).
The regulations C1->C2, C2->C3 and C3->C1 are favored.
}
}
