% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNRank.R
\name{RNRank}
\alias{RNRank}
\title{Rank by gene importance}
\usage{
RNRank(
  srcm,
  damping = 1,
  max_iterations = 100,
  threshold = 0,
  self = F,
  letZeros = F,
  divider = 100,
  sorted = T,
  verbose = F,
  throwOnError = T
)
}
\arguments{
\item{srcm}{Πίνακας ονομάτων γονιδίων 2 στηλών. Κάθε γραμμή περιέχει ένα ρυθμίζον
κι ένα ρυθμιζόμενο γονίδιο.}

\item{damping}{Συντελεστής απόσβεσης (0-1) που καθορίζει ποσοστό μη τυχαιότητας.}

\item{max_iterations}{Ο μέγιστος αριθμός επαναλήψεων, αν δεν επιτευχθεί σύγκλιση νωρίτερα (minimum 10).}

\item{threshold}{Ευκλίδεια απόσταση μικρότερη ή ίση από αυτήν την τιμή μεταξύ των βημάτων,
θα τερματίσει την επανάληψη.}

\item{self}{Καθορίζει αν επιτρέπονται οι εγγραφές αυτορρύθμισης.}

\item{letZeros}{Καθορίζει αν επιτρέπουμε τυχαίες μεταβάσεις από τα ρυθμίζοντα γονίδια
προς άλλα που δεν καθορίζονται από τον πίνακά μας.}

\item{divider}{Καθορίζει διαιρέτη του ποσοστού που υπολογίζεται για τις μεταβάσεις
από τα μη ρυθμίζοντα γονίδια, ώστε το αποτέλεσμα της διαίρεσης να χρησιμοποιηθεί
για τις τυχαίες/άγνωστες μεταβάσεις από τα ρυθμίζοντα (minimum 10).}

\item{sorted}{Καθορίζει αν τα αποτελέσματα θα επιστραφούν ταξινομημένα κατά ποσοστό.}

\item{verbose}{Εμφανίζει κάποια μηνύματα κατά τη διαδικασία.}

\item{throwOnError}{Αν είναι FALSE, σε περίπτωση σοβαρού λάθους επιστρέφει NULL.
Αλλιώς καλείται η stop().}
}
\value{
Ονοματισμένος πίνακας ποσοστών (0-1) 1 στήλης.
}
\description{
Υπολογισμός σημαντικότητας γονιδίων ρυθμιστικού δικτύου με χρήση αλγορίθμου
παρόμοιου με τον PageRank της Google.
}
